Moved utilities to opm/utility/graph and opm/utility/numeric.
Also fixed whitespace in CMakeFiles_list.cmake.
