+++ opm/utility/graph/AssembledConnections.hpp
