+++ opm/utility/numeric/RandomVector.cpp
@@ -22,7 +22,7 @@
 #include <config.h>
 #endif // HAVE_CONFIG_H
 
-#include <opm/flowdiagnostics/utility/RandomVector.hpp>
+#include <opm/utility/numeric/RandomVector.hpp>
 
 #include <random>
